refactor: append_to_labels
